+++ django_project/geosight/data/tests/django_admin/_base.py
@@ -18,8 +18,10 @@ import copy
 
 from django.contrib.auth import get_user_model
 from django.urls import reverse
+from rest_framework.reverse import reverse
 
-from core.tests.model_factories import UserF
+from core.models.profile import ROLES
+from core.tests.model_factories import UserF, create_user
 from geosight.permission.tests._base import BasePermissionTest
 
 User = get_user_model()
@@ -109,3 +111,20 @@ class BaseDjangoAdminTest(object):
             self.assertEqual(
                 self.resource.modified_by_username, self.admin.username
             )
+
+        def test_user_deleted(self):
+            """Test user deleted."""
+            user = create_user(ROLES.CREATOR.name)
+            resource = self.create_resource(user)
+            self.assertEqual(resource.creator, user)
+            self.assertEqual(resource.creator_username, user.username)
+            self.assertEqual(resource.modified_by, user)
+            self.assertEqual(resource.modified_by_username, user.username)
+
+            # Delete user
+            user.delete()
+            resource.refresh_from_db()
+            self.assertEqual(resource.creator, None)
+            self.assertEqual(resource.creator_username, user.username)
+            self.assertEqual(resource.modified_by, None)
+            self.assertEqual(resource.modified_by_username, user.username)
